+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Changed
-- Changed the output filename for RSV from `reg_dn.csv` to `reserves.csv`.
-- Created link in the thermal doc page to fuels page for piecewise linear cost curve documentation
+- Changed the output filename for RSV from `reg_dn.csv` to `reserves.csv` (#814).
+- Created link in the thermal doc page to fuels page for piecewise linear cost curve documentation (#820).
+- Doc rendering fix in fuel.jl (#821).
 
 ### Fixed
 - Fixed capacity reserve margin formulation for asymmetric storage